refactor: extract hardware-key capture/repeat into TerminalInputTextView+HardwareKeys; PR template + docstrings
Resolve the "Swift File And Package Boundaries" finding by genuinely shrinking
TerminalInputTextView.swift instead of growing it.

- New TerminalInputTextView+HardwareKeys.swift: the thin pressesBegan/Ended/
  Cancelled + keyCommands + resign/becomeFirstResponder + paste overrides that
  adapt UIKit press batches to the already-extracted TerminalHardwareKeyCapture
  (the capture/repeat/encoding logic lives there). Stored props stay on the view.
- New TerminalInputTextView+UITextInput.swift: the UITextInputTraits +
  documentless UITextInput conformance and the zero-width delete-repeat anchor /
  virtual-document model that conformance consumes.
- TerminalInputTextView.swift: 1581 -> 1322 lines (net -259, a real reduction,
  not +growth). A few private members widened to internal so the same-target
  extensions compile.
- swift-file-length-budget.tsv: reduce this file's entry 1629 -> 1322.
- Docstrings added to the new overrides and other undocumented new declarations.

Verified: cmux-ios build SUCCEEDED (iOS Simulator); CmuxMobileTerminalKit
swift test (96 tests); CmuxMobileTerminal sim tests (22 tests: hardware-key
capture + hold-to-repeat + selection); CmuxTerminal macOS tests (56 tests).
Hardware-key capture and hold-to-repeat behavior unchanged.
