+++ Packages/iOS/CmuxMobileTerminal/Sources/CmuxMobileTerminal/TerminalInputTextView.swift
@@ -89,15 +89,19 @@ final class TerminalInputTextView: UIView, UIKeyInput, UITextInput {
     /// can answer the framework's read-back. On ``unmarkText()`` the held string
     /// commits through ``insertText(_:)`` and this clears. There is no
     /// committed-document buffer.
-    private var markedText: String?
+    /// The transient IME composition string (the only text this view holds);
+    /// read by the ``UITextInput`` conformance in
+    /// `TerminalInputTextView+UITextInput.swift`, hence package-internal.
+    var markedText: String?
 
     /// Toggles the zero-width character used for the delete-repeat anchor on each
     /// empty-buffer ``deleteBackward()``. Alternating between two distinct
     /// zero-width characters guarantees the virtual document's *contents* change
     /// every delete, which (alongside the `inputDelegate` text-change brackets)
     /// re-arms UIKit's document-driven backspace repeat. Borrowed verbatim from
-    /// vvterm's `deleteRepeatAnchorUsesAlternate`.
-    private var deleteRepeatAnchorUsesAlternate = false
+    /// vvterm's `deleteRepeatAnchorUsesAlternate`. Internal so the virtual-document
+    /// model in `TerminalInputTextView+UITextInput.swift` can toggle it.
+    var deleteRepeatAnchorUsesAlternate = false
 
     /// Monotonic-ish tap timestamp for the reducer's double-tap window. Uses
     /// the same wall-clock source the legacy `Date()` comparisons used, so the
@@ -129,161 +133,20 @@ final class TerminalInputTextView: UIView, UIKeyInput, UITextInput {
         set {}
     }
 
-    override var keyCommands: [UIKeyCommand]? {
-        TerminalInputDebugLog.log("proxy.keyCommands.getter")
-        guard markedText == nil else { return nil }
-        // Hardware keys are captured exclusively by `pressesBegan` (which runs
-        // below the text-editing layer), so there is no nav/Control/Shift-Arrow
-        // `UIKeyCommand` table here — a second `UIKeyCommand` handler racing
-        // `pressesBegan` for the same chords was the source of the inconsistent
-        // ("janky") modifier behavior. The lone surviving command is Cmd+V:
-        // `shouldConsume` deliberately lets plain Command fall through, so paste
-        // reaches this routing (a bare UIView does not inherit UITextView's
-        // Cmd+V). Claim priority so the text-input layer cannot eat it first.
-        let pasteCommand = UIKeyCommand(input: "v", modifierFlags: [.command], action: #selector(paste(_:)))
-        pasteCommand.wantsPriorityOverSystemBehavior = true
-        return [pasteCommand]
-    }
-
-    // MARK: Hardware-keyboard capture (pressesBegan — below the text system)
-    //
-    // The direct-terminal first responder is this `UIKeyInput`/`UITextInput`
-    // proxy. With UIKit's text-editing layer focused on the zero-width virtual
-    // document, arrows/Tab/Ctrl-nav are consumed as no-op caret edits BEFORE
-    // `keyCommands` (or `wantsPriorityOverSystemBehavior`) can fire. Capturing
-    // at `pressesBegan` runs below the text system, so special keys and
-    // Control/Option chords reach the terminal; everything else falls through to
-    // `super` so normal typing/IME/dictation still routes via `insertText`. The
-    // capture decision, byte encoding, and timer-driven hold-to-repeat all live
-    // in the dedicated `TerminalHardwareKeyCapture`; the overrides below are thin
-    // adapters that hand UIKit's press batches to it and forward the remainder to
-    // `super`.
-
     /// Owns hardware-key capture and timer-driven hold-to-repeat (see
     /// ``TerminalHardwareKeyCapture``), extracted so this view keeps only thin
-    /// press overrides. `isComposing` blocks capture during IME marking; `emit`
-    /// routes encoded bytes to the ``onEscapeSequence`` send sink.
-    private lazy var hardwareKeyCapture = TerminalHardwareKeyCapture(
+    /// press overrides (which live in `TerminalInputTextView+HardwareKeys.swift`).
+    /// `isComposing` blocks capture during IME marking; `emit` routes encoded
+    /// bytes to the ``onEscapeSequence`` send sink.
+    lazy var hardwareKeyCapture = TerminalHardwareKeyCapture(
         timerFactory: keyRepeatTimerFactory,
         isComposing: { [weak self] in self?.markedText != nil },
         emit: { [weak self] data in self?.onEscapeSequence?(data) }
     )
 
-    override func becomeFirstResponder() -> Bool {
-        let ok = super.becomeFirstResponder()
-        TerminalInputDebugLog.log("proxy.becomeFR cls=\(String(describing: type(of: self))) ok=\(ok)")
-        return ok
-    }
-
-    override func resignFirstResponder() -> Bool {
-        // A hardware key held while focus leaves this proxy never delivers its
-        // `pressesEnded`, so cancel every in-flight key-repeat (and drop the
-        // captured-press set whose matching ends will never arrive) before
-        // yielding first responder.
-        hardwareKeyCapture.reset()
-        return super.resignFirstResponder()
-    }
-
-    override func pressesBegan(_ presses: Set<UIPress>, with event: UIPressesEvent?) {
-        let forwarded = hardwareKeyCapture.pressesBegan(presses)
-        if !forwarded.isEmpty { super.pressesBegan(forwarded, with: event) }
-    }
-
-    override func pressesEnded(_ presses: Set<UIPress>, with event: UIPressesEvent?) {
-        let forwarded = hardwareKeyCapture.pressesEnded(presses)
-        if !forwarded.isEmpty { super.pressesEnded(forwarded, with: event) }
-    }
-
-    override func pressesCancelled(_ presses: Set<UIPress>, with event: UIPressesEvent?) {
-        let forwarded = hardwareKeyCapture.pressesCancelled(presses)
-        if !forwarded.isEmpty { super.pressesCancelled(forwarded, with: event) }
-    }
-
-    /// Restores standard system paste on this documentless responder.
-    ///
-    /// As a `UITextView` the view inherited `paste(_:)`/`canPerformAction(_:)`;
-    /// as a bare `UIView` it must re-expose them so hardware Cmd+V and the
-    /// edit-menu Paste keep working. Only paste is advertised — copy/cut/select
-    /// are meaningless on a proxy that holds no document, so they stay disabled
-    /// rather than surfacing a broken edit menu.
-    override func canPerformAction(_ action: Selector, withSender sender: Any?) -> Bool {
-        if action == #selector(paste(_:)) {
-            return UIPasteboard.general.hasStrings || UIPasteboard.general.hasImages
-        }
-        return false
-    }
-
-    /// System Paste (Cmd+V or the edit-menu item) routed through the same
-    /// clipboard handling as the toolbar Paste button: an image goes to the Mac
-    /// as `terminal.paste_image`, text rides the normal input sink.
-    override func paste(_ sender: Any?) {
-        handlePasteAction()
-    }
-
-    // MARK: Hold-to-repeat backspace (hasText + zero-width anchor)
-
-    /// Always report that there is text to delete.
-    ///
-    /// This is the legacy ``UIKeyInput`` gate (borrowed from iSH's
-    /// `TerminalView`) for the software keyboard's *hold-to-repeat* backspace. On
-    /// a bare ``UIKeyInput`` responder (this view) the keyboard's auto-repeat
-    /// timer keeps firing ``deleteBackward()`` only while the first responder
-    /// reports `hasText == true`. It is always safe to send a DEL byte to the
-    /// remote terminal, so there is no "nothing to delete" state to honor —
-    /// return `true` unconditionally.
-    ///
-    /// Internal byte-routing therefore must *not* key off `hasText` (it is a
-    /// constant); ``deleteBackward()`` and the modifier guards key off
-    /// ``markedText`` (IME composition) instead.
-    var hasText: Bool { true }
-
-    /// Whether the view is currently presenting the zero-width delete-repeat
-    /// anchor: there is no IME composition, so the view fabricates a
-    /// one-character virtual document with the caret at the end. While this is
-    /// true UIKit's modern document-driven repeat path sees a deletable character
-    /// to the left of the cursor and keeps firing ``deleteBackward()``. Mirrors
-    /// vvterm's `usesDeleteRepeatAnchor`.
-    private var usesDeleteRepeatAnchor: Bool { markedText == nil }
-
-    /// The current zero-width anchor character. Toggling between two distinct
-    /// zero-width characters on each empty-buffer delete forces the virtual
-    /// document's contents to change, which is what re-arms the repeat. (vvterm:
-    /// `deleteRepeatAnchorText`.)
-    private var deleteRepeatAnchorText: String {
-        deleteRepeatAnchorUsesAlternate ? "\u{2060}" : "\u{200B}"
-    }
-
-    /// The virtual document UIKit walks: the IME composition while marking,
-    /// otherwise the one-character zero-width delete-repeat anchor.
-    private var textInputDocument: String {
-        usesDeleteRepeatAnchor ? deleteRepeatAnchorText : (markedText ?? "")
-    }
-
-    private var textInputDocumentLength: Int {
-        (textInputDocument as NSString).length
-    }
-
-    /// The caret sits at the end of the virtual document (after the anchor char,
-    /// or at the end of the marked composition) so there is always something to
-    /// its left to delete.
-    private var effectiveSelectedRange: NSRange {
-        NSRange(location: textInputDocumentLength, length: 0)
-    }
-
-    /// Re-arm the delete-repeat anchor after an empty-buffer delete.
-    ///
-    /// Toggling the anchor char inside `textWillChange`/`textDidChange` brackets
-    /// tells UIKit the (still one-character) virtual document changed, so the
-    /// keyboard's document-driven repeat timer re-reads it and fires the next
-    /// ``deleteBackward()``. Without this the modern repeat path stalls after one
-    /// delete even though ``hasText`` stays `true` — this is the refinement the
-    /// prior documentless attempt lacked. Mirrors vvterm's
-    /// `notifyVirtualDeleteAnchorDidChange`.
-    private func notifyVirtualDeleteAnchorDidChange() {
-        inputDelegate?.textWillChange(self)
-        deleteRepeatAnchorUsesAlternate.toggle()
-        inputDelegate?.textDidChange(self)
-    }
+    // The `hasText` gate and the zero-width delete-repeat anchor model (the
+    // virtual document UIKit walks) live alongside the `UITextInput` conformance
+    // that consumes them, in `TerminalInputTextView+UITextInput.swift`.
 
     private static let monokaiBarColor = UIColor(red: 0x27/255.0, green: 0x28/255.0, blue: 0x22/255.0, alpha: 1)
     private static let accessoryHorizontalInset: CGFloat = 16
@@ -1180,7 +1043,11 @@ final class TerminalInputTextView: UIView, UIKeyInput, UITextInput {
     /// falls back to JPEG so it stays under the Mac's 10 MB cap. Accessing the
     /// pasteboard contents here is what shows iOS's one-shot paste banner, which
     /// is the expected confirmation for an explicit Paste tap.
-    private func handlePasteAction() {
+    /// Routes clipboard contents into the terminal: an image is sent to the Mac
+    /// as `terminal.paste_image`, text rides the normal input sink. Internal so
+    /// the `paste(_:)` override in `TerminalInputTextView+HardwareKeys.swift` can
+    /// share this path with the toolbar Paste button.
+    func handlePasteAction() {
         let pasteboard = UIPasteboard.general
         if pasteboard.hasImages, let image = pasteboard.image {
             let maxImageBytes = 8 * 1024 * 1024
@@ -1243,7 +1110,10 @@ final class TerminalInputTextView: UIView, UIKeyInput, UITextInput {
         }
     }
 
-    private func emitCommittedText(_ committedText: String, source: String) {
+    /// Applies any armed modifiers and forwards committed text to the terminal.
+    /// Internal so the ``unmarkText()``/``replace(_:withText:)`` commit paths in
+    /// `TerminalInputTextView+UITextInput.swift` share this single send path.
+    func emitCommittedText(_ committedText: String, source: String) {
         TerminalInputDebugLog.log("proxy.emit source=\(source) len=\(committedText.count)")
         if controlAccessoryArmed {
             if !controlAccessorySticky {
@@ -1447,183 +1317,6 @@ final class TerminalInputTextView: UIView, UIKeyInput, UITextInput {
     #endif
 }
 
-// MARK: - UITextInputTraits
-
-extension TerminalInputTextView {
-    // Autocorrect/predictive/smart substitutions are all off: the view forwards
-    // each keystroke to the remote terminal and keeps no in-progress word for the
-    // keyboard to correct against. Returning these as computed properties (rather
-    // than the `UITextView` stored traits the old design used) keeps the keyboard
-    // from offering corrections it could never apply.
-    var autocorrectionType: UITextAutocorrectionType { get { .no } set {} }
-    var autocapitalizationType: UITextAutocapitalizationType { get { .none } set {} }
-    var spellCheckingType: UITextSpellCheckingType { get { .no } set {} }
-    var smartQuotesType: UITextSmartQuotesType { get { .no } set {} }
-    var smartDashesType: UITextSmartDashesType { get { .no } set {} }
-    var smartInsertDeleteType: UITextSmartInsertDeleteType { get { .no } set {} }
-    var keyboardType: UIKeyboardType { get { .default } set {} }
-    var returnKeyType: UIReturnKeyType { get { .default } set {} }
-}
-
-// MARK: - UITextInput (documentless conformance + delete-repeat anchor)
-
-// This view owns no editable document. It implements `UITextInput` to unlock two
-// keyboard features a bare `UIKeyInput` view does not get — system dictation (the
-// mic key) and IME marked-text composition — and to present vvterm's toggling
-// one-character zero-width *delete-repeat anchor* so the software keyboard's
-// modern document-driven backspace auto-repeat keeps firing while the key is
-// held. Recognized dictation text and committed IME candidates both arrive
-// through ``insertText(_:)`` and route to the terminal. UIKit walks the virtual
-// document (`textInputDocument`) through the offset-bearing
-// ``TerminalInputTextPosition`` / ``TerminalInputTextRange`` types, so positions
-// and ranges carry real UTF-16 offsets rather than acting as pure identity
-// sentinels — that is what lets the view report a one-character document with the
-// caret at the end, the condition the repeat path checks.
-extension TerminalInputTextView {
-    var markedTextRange: UITextRange? {
-        guard let markedText, !markedText.isEmpty else { return nil }
-        return TerminalInputTextRange(start: 0, end: (markedText as NSString).length)
-    }
-
-    var selectedTextRange: UITextRange? {
-        get {
-            let range = effectiveSelectedRange
-            return TerminalInputTextRange(start: range.location, end: range.location + range.length)
-        }
-        set {}
-    }
-
-    var markedTextStyle: [NSAttributedString.Key: Any]? {
-        get { nil }
-        set {}
-    }
-
-    var beginningOfDocument: UITextPosition { TerminalInputTextPosition(offset: 0) }
-    var endOfDocument: UITextPosition { TerminalInputTextPosition(offset: textInputDocumentLength) }
-
-    /// The IME hands a candidate string in; hold it as the marked composition so
-    /// ``markedTextRange`` reports active composition (which also suppresses the
-    /// delete-repeat anchor). Nothing is sent to the terminal until the candidate
-    /// commits via ``insertText(_:)`` or ``unmarkText()``.
-    ///
-    /// Mutating ``markedText`` changes the string the view exposes through
-    /// ``text(in:)``/``markedTextRange``, so it is a *text* change in the
-    /// ``UITextInputDelegate`` contract: it is bracketed with
-    /// `textWillChange`/`textDidChange` (via ``withMarkedTextChange(_:)``) so the
-    /// IME and dictation machinery keep their composition state synchronized.
-    func setMarkedText(_ markedText: String?, selectedRange: NSRange) {
-        TerminalInputDebugLog.log("proxy.setMarkedText len=\((markedText ?? "").count)")
-        withMarkedTextChange {
-            self.markedText = (markedText?.isEmpty == true) ? nil : markedText
-        }
-    }
-
-    /// Brackets a mutation of ``markedText`` (or the anchor crossover when
-    /// composition ends) with the `UITextInputDelegate` text-change callbacks.
-    ///
-    /// The marked composition is the only committed text this view exposes, so
-    /// any change to it — set by the IME, committed by
-    /// ``insertText(_:)``/``unmarkText()``, or canceled by ``deleteBackward()`` —
-    /// is a text change UIKit must be told about with
-    /// `textWillChange`/`textDidChange`. Selection-only callbacks would leave the
-    /// keyboard observing stale composition state.
-    private func withMarkedTextChange(_ mutate: () -> Void) {
-        inputDelegate?.textWillChange(self)
-        mutate()
-        inputDelegate?.textDidChange(self)
-    }
-
-    /// Commit the in-progress IME composition. Forwards the held candidate to the
-    /// terminal as one block.
-    func unmarkText() {
-        guard let composing = markedText else { return }
-        withMarkedTextChange { markedText = nil }
-        emitCommittedText(composing, source: "unmarkText")
-    }
-
-    func text(in range: UITextRange) -> String? {
-        guard let range = range as? TerminalInputTextRange else { return nil }
-        let document = textInputDocument as NSString
-        let clamped = clampedDocumentRange(range.nsRange, length: document.length)
-        guard clamped.length > 0 else { return "" }
-        return document.substring(with: clamped)
-    }
-
-    /// Commit text delivered through a range replacement.
-    ///
-    /// Most committed input arrives via ``insertText(_:)``, but some system paths
-    /// (text replacement, certain dictation/suggestion commits) deliver it by
-    /// replacing ``selectedTextRange`` or ``markedTextRange`` instead. The view
-    /// holds no addressable document, so the range itself is ignored, but the
-    /// *text* must still reach the terminal — route it through the same commit
-    /// path as ``insertText(_:)``. A replacement of the marked region supersedes
-    /// the in-progress IME composition, so clear it first. An empty replacement is
-    /// a pure deletion of the marked composition (no committed text to send).
-    func replace(_ range: UITextRange, withText text: String) {
-        TerminalInputDebugLog.log("proxy.replace len=\(text.count)")
-        if markedText != nil {
-            withMarkedTextChange { markedText = nil }
-        }
-        guard !text.isEmpty else { return }
-        emitCommittedText(text, source: "replace")
-    }
-
-    func textRange(from fromPosition: UITextPosition, to toPosition: UITextPosition) -> UITextRange? {
-        guard let from = fromPosition as? TerminalInputTextPosition,
-              let to = toPosition as? TerminalInputTextPosition else { return nil }
-        return TerminalInputTextRange(start: from.offset, end: to.offset)
-    }
-
-    func position(from position: UITextPosition, offset: Int) -> UITextPosition? {
-        guard let position = position as? TerminalInputTextPosition else { return nil }
-        let target = min(max(position.offset + offset, 0), textInputDocumentLength)
-        return TerminalInputTextPosition(offset: target)
-    }
-
-    func position(from position: UITextPosition, in direction: UITextLayoutDirection, offset: Int) -> UITextPosition? {
-        guard let position = position as? TerminalInputTextPosition else { return nil }
-        let delta = (direction == .right || direction == .down) ? offset : -offset
-        let target = min(max(position.offset + delta, 0), textInputDocumentLength)
-        return TerminalInputTextPosition(offset: target)
-    }
-
-    func compare(_ position: UITextPosition, to other: UITextPosition) -> ComparisonResult {
-        guard let lhs = position as? TerminalInputTextPosition,
-              let rhs = other as? TerminalInputTextPosition else { return .orderedSame }
-        if lhs.offset < rhs.offset { return .orderedAscending }
-        if lhs.offset > rhs.offset { return .orderedDescending }
-        return .orderedSame
-    }
-
-    func offset(from: UITextPosition, to toPosition: UITextPosition) -> Int {
-        guard let from = from as? TerminalInputTextPosition,
-              let to = toPosition as? TerminalInputTextPosition else { return 0 }
-        return to.offset - from.offset
-    }
-
-    func position(within range: UITextRange, farthestIn direction: UITextLayoutDirection) -> UITextPosition? { nil }
-    func characterRange(byExtending position: UITextPosition, in direction: UITextLayoutDirection) -> UITextRange? { nil }
-    func baseWritingDirection(for position: UITextPosition, in direction: UITextStorageDirection) -> NSWritingDirection { .leftToRight }
-    func setBaseWritingDirection(_ writingDirection: NSWritingDirection, for range: UITextRange) {}
-    func firstRect(for range: UITextRange) -> CGRect { .zero }
-    func caretRect(for position: UITextPosition) -> CGRect { .zero }
-    func selectionRects(for range: UITextRange) -> [UITextSelectionRect] { [] }
-    func closestPosition(to point: CGPoint) -> UITextPosition? { nil }
-    func closestPosition(to point: CGPoint, within range: UITextRange) -> UITextPosition? { nil }
-    func characterRange(at point: CGPoint) -> UITextRange? { nil }
-
-    private func clampedDocumentRange(_ range: NSRange, length: Int) -> NSRange {
-        let location = min(max(range.location, 0), length)
-        let rangeLength = min(max(range.length, 0), max(length - location, 0))
-        return NSRange(location: location, length: rangeLength)
-    }
-
-    // MARK: Dictation placeholder hooks
-    //
-    // UIKit calls these when the mic is tapped. Returning a placeholder (an empty
-    // token; iSH does the same) is what tells the framework this view accepts
-    // dictation; the recognized text then arrives via `insertText`. The remove
-    // hook is a no-op because there is no document placeholder to strip.
-    func insertDictationResultPlaceholder() -> Any { "" }
-    func removeDictationResultPlaceholder(_ placeholder: Any, willInsertResult: Bool) {}
-}
+// UITextInputTraits and the documentless UITextInput conformance (plus the
+// zero-width delete-repeat anchor's text(in:)/marked-text plumbing) live in
+// `TerminalInputTextView+UITextInput.swift`.
